Tried fixing move and zoom logic with D2D1 drawing

+++ NodePlanningEditor/src/NPE/Handlers/Mouse.cpp
@@ -91,7 +91,7 @@ namespace NPE
 
 	}
 
-	void Mouse::OnMouseMove(float x, float y)
+	void Mouse::OnMouseMove(int x, int y)
 	{
 		m_Pos = NPoint{x, y};
 		m_Queue.push(Event(Event::Type::Move, this));
